bbq_manager/requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
dill==0.3.5.1
Django==4.1
django-cors-headers==3.13.0
django-filter==22.1
django-service-objects==0.7.1
djangorestframework==3.13.1
drf-yasg==1.21.3
Faker==12.0.1
flake8==5.0.4
funcy==1.17
greenlet==1.1.2
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
mixer==7.2.2
msgpack==1.0.4
packaging==21.3
platformdirs==2.5.2
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.9.1
pyflakes==2.5.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tomlkit==0.11.4
uritemplate==4.1.1
urllib3==1.26.11
wrapt==1.14.1
gunicorn